+++ src/structurizr/model/model_item.py
@@ -84,7 +84,17 @@ class ModelItem(AbstractBase, ABC):
         **kwargs,
     ):
         """Initialise a ModelItem instance."""
-        super().__init__(**kwargs)
+        if len(kwargs) > 0:
+            type_name = type(self).__name__
+            args = [f"'{key}'" for key in kwargs.keys()]
+            if len(args) == 1:
+                raise TypeError(f"{type_name}.__init__() got an unexpected "
+                                f"keyword argument {args[0]}")
+            else:
+                raise TypeError(f"{type_name}.__init__() got unexpected "
+                                f"keyword arguments {', '.join(args)}")
+
+        super().__init__()
         self.id = id
         self.tags = OrderedSet(tags)
         self.properties = dict(properties)